Fix&improve RPC implementation
Catch errors of the function itself, not just of its returned promise.

Throw two types of error, do distinguish between failure *to call* the
function and failure *of* the function.

Ensure the handler always returns a Promise; otherwise it will not be
considered a return message (at least in webextension-polyfill).